requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlmodel==0.0.22
pydantic==2.8.2
pydantic-settings==2.2.1
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
requests==2.32.3
jinja2==3.1.4
